back/requirements/base.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-environ==0.4.5
django-extensions==2.1.9
django-model-utils==3.2.0
djangorestframework==3.10.1
idna==2.8
oauthlib==3.0.2
psycopg2==2.8.3
Pygments==1.6
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlformatter==1.3
sqlparse==0.1.11
urllib3==1.25.3
